Close slave fd when on UNIX + normalize interfaces
Sockets in NodeJS usually buffer as much data as possible. When opening
a pty, node-pty wraps both the master fd and the slave fd in sockets.

Issue is that we then get a stream loop, where things on the slave side
might be starved of input because of the wrapping done.

We actually don't need the slave fd because we should write in master.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>